Update README.md
This notebook is intended to be used for photobleach calculations based on a kymograph collection of static fluorophores.

To setup a collection, begin by performing continuous confocal scans while lowering the Z-position with the silver knob. When the Z focus is at the bottom edge of the slide, proteins stuck to the surface should fluoresce brightly. To analyze photobleaching, start collecting a kymograph as in data collection settings until photon counts decrease to less than half of the original value. This script should automatically fetch and bin the photon counts in the blue channel by second and plot them as a function of time. Finally, an exponential decay fit determines the lifetime of the fluorophore (see figure 2) Note that color still must be changed in cells 5 and 7 for colors other than blue